Task #115: Fix sweeper fake durations & offset_ms int4 overflow
Two bugs in complete_call() when invoked by the stuck-call sweeper, both
triggered by very old stuck calls (answered_at IS NULL).

Bug 1 — 800-minute fabricated durations (call_logger.py):
  Root cause: leg.ended_at = datetime.utcnow() and leg.duration_seconds =
  (now - leg.started_at). A call ringing at 15:55 swept at 05:59 next day
  recorded 840 fake minutes. Nine production rows had 804–28,333 min durations.
  Fix: _skip_sweeper_duration flag (forced_by=="sweeper" and not answered_at).
  When set, skip duration computation on both call_log and all legs.
  leg.ended_at is still stamped to close the record. Leg duration stays 0.

Bug 2 — NumericValueOutOfRange, two calls stuck forever (call_logger.py):
  Root cause: _write_event_inline computed offset_ms = (now - started_at) * 1000.
  For calls stuck >24.8 days (int4 max = 2,147,483,647 ms), this overflows
  PostgreSQL INTEGER. The commit rolls back; sweeper retries every 5 min forever.
  Fix A: cap offset_ms at min(raw_ms, _INT4_MAX) in _write_event_inline.
  Fix B: ALTER TABLE call_events ALTER COLUMN offset_ms TYPE BIGINT (migration).
  Fix C: CallEvent model updated from Integer to BigInteger.

Schema migration (database.py _ADDITIVE_MIGRATIONS):
  - Idempotent DO block: widens call_events.offset_ms from int4 to int8 if still narrow.
  - Idempotent UPDATE: zeros call_logs.duration_seconds where answered_at IS NULL
    and duration_seconds > 7200 (conservative guard, covers all 9 bad prod rows).
  Both run on the next production startup (deploy needed to apply to prod).

Tests (tests/test_sweeper_complete_call.py, 7 new cases):
  - Unanswered sweeper call: duration stays 0 (not fabricated from started_at)
  - leg.ended_at/call_log.ended_at still stamped to close record
  - 25-day-old call finalizes without overflow (result=True, duration=0)
  - Unanswered call classified as ended_early
  - Answered sweeper call still computes duration correctly (not regressed)
  - offset_ms capped at _INT4_MAX for all written events
  - _INT4_MAX sanity check (2^31 - 1 = 2,147,483,647)

All 25 tests pass. Production fix takes effect on next deploy.

+++ botelier/backend/botelier/models/call_event.py
@@ -9,7 +9,7 @@ root-cause diagnosis. Events are written in two ways:
 
 import uuid
 from datetime import datetime
-from sqlalchemy import Column, String, DateTime, Integer, Index, ForeignKey
+from sqlalchemy import Column, String, DateTime, BigInteger, Index, ForeignKey
 from sqlalchemy.dialects.postgresql import UUID, JSONB
 from botelier.database import Base
 
@@ -44,7 +44,7 @@ class CallEvent(Base):
     severity = Column(String, nullable=False, default="info")
 
     occurred_at = Column(DateTime, nullable=False, default=datetime.utcnow)
-    offset_ms = Column(Integer, nullable=True)
+    offset_ms = Column(BigInteger, nullable=True)
 
     details = Column(JSONB, nullable=True)
